[local] ctm2tg: ensure causality between curr bos and prev eos (#16)
