+++ requirements.txt
@@ -1,6 +1,5 @@
 gdown==5.0.0
 numpy==1.26.3
-pandas==2.2.0
-#tqdm==4.66.1
+pandas==2.2.2
 TextGrid==1.5
 Unidecode==1.3.8
